pyproject.toml
[project]
name = "knowledge-forge"
version = "0.1.0"
description = "企业级智能知识库问答系统 - RAG + MCP Server + Agent"
readme = "README.md"
requires-python = ">=3.11,!=3.12.0,!=3.12.1,!=3.12.2"
dependencies = [
    # Web 框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.0.0",

    # 数据验证 & 配置
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # 文档解析
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "mistune>=3.0.0",

    # 向量数据库
    "pymilvus>=2.5.0",

    # 关系数据库
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Redis
    "redis>=5.0.0",

    # Embedding & LLM
    "openai>=1.60.0",
    "tiktoken>=0.9.0",

    # Reranker
    "FlagEmbedding>=1.2.0",

    # BM25
    "rank-bm25>=0.2.2",
    "jieba>=0.42.1",

    # MCP
    "mcp>=1.0.0",

    # 异步任务
    "celery>=5.4.0",

    # 评估 & 可视化
    "streamlit>=1.40.0",
    "plotly>=5.24.0",

    # 工具
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
knowledge-forge = "knowledge_forge.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_forge"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
